Engine: hide "loopcounter" variable under getter/setter

+++ Engine/ac/gamestate.h
@@ -240,7 +240,7 @@ struct GamePlayState
     int   silent_midi = 0;
     int   silent_midi_channel = 0;
     int   current_music_repeating = 0;  // remember what the loop flag was when this music started
-    unsigned long shakesc_delay = 0;  // unsigned long to match loopcounter
+    uint32_t shakesc_delay = 0;  // unsigned long to match loopcounter
     int   shakesc_amount = 0;
     int   shakesc_length = 0;
     int   rtint_red = 0;
